Added Login and Logout

+++ MovieRecommender/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib import messages
+from django.contrib.auth.models import User
 from .models import Room, Topic
 from .forms import RoomForm
-
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.http import HttpResponse
 
 rooms = [
     {'id': 1, 'name': 'Lets learn python!'},
@@ -12,7 +16,7 @@ rooms = [
 
 
 def home(request):
-    q= request.GET.get('q') if request.GET.get('q') is not None else ''
+    q = request.GET.get('q') if request.GET.get('q') is not None else ''
     rooms = Room.objects.filter(
         Q(topic__name__contains=q) |
         Q(name__icontains=q) |
@@ -33,7 +37,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'MovieRecommender/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -50,6 +54,9 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -60,9 +67,41 @@ def updateRoom(request, pk):
     return render(request, 'MovieRecommender/room_form.html', context)
 
 
-def deleteRoom(request,pk):
+def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
-    if request.method =='POST':
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
+    if request.method == 'POST':
         room.delete()
         return redirect('home')
-    return render(request, 'MovieRecommender/delete.html', {'obj':room})
+    return render(request, 'MovieRecommender/delete.html', {'obj': room})
+
+
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist')
+
+    context = {}
+    return render(request, 'MovieRecommender/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
